fix: rename folder/files name
